requirements.txt
dataclasses;python_version<'3.7'
python-nubia
jellyfish
prettytable
prompt-toolkit>=2
Pygments
pyparsing>=2.2.0
termcolor
wcwidth
web3
requests